reorganize and rename
