+++ template-single-doc/release-artifact-template.yaml
@@ -1,0 +1,11 @@
+# --- Release notes artifact ----
+
+charm_name: ""
+release_condition: ""
+release_tag: ""
+changes:
+  - title: ""
+    author: ""
+    type: ""
+    description: ""
+    url: ""
