+++ template-single-doc/release-template.md.j2
@@ -30,33 +30,26 @@ Use this section to highlight major and minor features that were added in this r
 The subsection below shows the pattern for each feature. Include links to the relevant PR or commit.
 -->
 
-{% for feat in change -%}
-{%- if feat.type == "major" or feat.type == "minor" -%}
+{% for feat in changes %}
+{% if feat.type == "major" or feat.type == "minor" %}
 ### {{ feat.title }}
 {{ feat.description }}
 <Add more context and information about the entry>
-
-Relevant links:
-* [PR]({{ feat.urls.pr }})
-{%- if feat.urls.related_doc|length -%}
-* [Related documentation]({{ feat.urls.related_doc }})
-{%- endif -%}
-{%- if feat.urls.related_issue|length -%}
-* [Related issue]({{ feat.urls.related_issue }})
-{%- endif %}
-{%- endif %}
+[Link to PR/commit]({{ feat.url }})
+{% endif %}
 {% endfor %}
+
 
 ## Bug fixes
 <!--
 Add a bulleted list of bug fixes here, with links to the relevant PR/commit.
 -->
-
-{% for feat in change -%}
-{%- if feat.type == "bugfix" -%}
-* {{ feat.title }} ([PR]({{ feat.urls.pr }})): {{ feat.description }}
-{% endif -%}
+{% for feat in changes %}
+{% if feat.type == "bugfix" %}
+* {{ feat.title }} ([PR]({{ feat.url }})): {{ feat.description }}
+{% endif %}
 {% endfor %}
+
 
 ## Breaking changes
 
@@ -66,21 +59,14 @@ Include links to the relevant PR or commit.
 If there are no breaking changes, keep the section and write "No breaking changes".
 -->
 
-{% for feat in change -%}
-{% if feat.type == "breaking" -%}
+{% for feat in changes %}
+{% if feat.type == "breaking" %}
 ### {{ feat.title }}
 {{ feat.description }}
 <Add more context and information about the entry>
+PR: {{ feat.url }}
 
-Relevant links:
-* [PR]({{ feat.urls.pr }})
-{%- if feat.urls.related_doc|length %}
-* [Related documentation]({{ feat.urls.related_doc }})
 {% endif %}
-{%- if feat.urls.related_issue|length %}
-* [Related issue]({{ feat.urls.related_issue }})
-{% endif -%}
-{% endif -%}
 {% endfor %}
 
 ## Deprecated
@@ -91,22 +77,15 @@ Include links to the relevant PR or commit.
 If there are no deprecated features, keep the section and write "No deprecated features".
 -->
 
-{% for feat in change -%}
-{% if feat.type == "deprecated" -%}
+{% for feat in changes %}
+{% if feat.type == "deprecated" %}
 ### {{ feat.title }}
 {{ feat.description }}
 <Add more context and information about the entry>
+PR: {{ feat.url }}
 
-Relevant links:
-* [PR]({{ feat.urls.pr }})
-{% if feat.urls.related_doc|length %}
-* [Related documentation]({{ feat.urls.related_doc }})
 {% endif %}
-{% if feat.urls.related_issue|length %}
-* [Related issue]({{ feat.urls.related_issue }})
-{% endif -%}
-{% endif -%}
-{% endfor -%}
+{% endfor %}
 
 ## Known issues
 <!--
@@ -121,8 +100,8 @@ If there are no known issues, keep the section and write "No known issues".
 <!--
 List of contributors based on PRs/commits. Remove this section if there are no contributors in this release.
 -->
-{% for feat in change -%}
-{{ feat.author }}, 
+{% for feat in changes %}
+* {{ feat.author }} 
 {% endfor %}
